Avoid heap test; die after failed alloc; sound++
Don't bother with _heapmaxavail(); just try to alloc(), and if
it fails, show an error and die.  (Gets the game running on my
real 1200XL again.)

Also, reduce volume of sounds, so they aren't so noisy.
Reset the note pointer when starting a new game (was left
as-is after a game-over event).