Add practice mode; overhaul README a little
Added a practice mode.  Press [P] (instead of [Start]) from
the title screen to begin a practice game.  When letters fill up,
the game does not end.  Your score is not eligible for recording
as a high score, however.

Press [Esc] during a practice game to end it (since it will not
end on its own).  This means [Esc] is no longer used to erase
your entire input word; only [Shift]+[Backspace] ("Delete") does
this now.

Also, removed some sections of README.md that do not apply to this
version of the game yet (regarding dictionary selection and
extending the game with new dictionaries).

Added details on the VBI-based sound routines, added yesterday,
to the "Some technical features" section of README.md.