Load lots of dicts onto disk; move eszett to [<]
Movin' on up to a double-density disk image (for now; may split
the game into separate ATRs, one for each language?).

Generating dictionaries for German, French, Italian, Spanish,
Polish, and British (vs American) English.  NOTE: No way to
pick any of them; still defaults to "EN_US.DIC".

Moved "ß" from "@" (`[Shift]`+`[8]` keys on Atari) symbol to the
"<" (`[<]` key on Atari) symbol.  Motivation: On German keyboards,
the "ß" symbol is on a key directly to the right of `[0]`; on the
Atari keyboard, the key directly to the right of `[0]` is `[<]`.