More work towards non-English dictionary support
Allow specifying number of letters when creating
dictionaries (as a way of keeping the size under control).

Reduce number of letters utilized with some languages
(French and Polish dictionaries on Linux are huge!)

Mostly figured out normalization of letters with diacritics
(eszett is not handled right, actually!)  Appearing possible
to play with other dictionaries now (more testing needed).

For debugging, export dump of all words for each dictionary
as a plaintext file.

Avoid binary search breaking when dictionary is not an
even number of words.  Also, when generating dictionaries,
sort the normalized results, to ensure binary search works
(e.g., "éclair" comes after "zygotes" in the American
English dictionary file, but we'd be storing it as [a mapping
of the usable letters] "ECLAIR", so it needs to come after
"EATS".)